change to generic shared weight manager
